+++ code/string_algorithms/src/arithmetic_on_large_numbers/string_factorial.cpp
@@ -22,7 +22,7 @@ int main(){
 		}
 	}
 	cout<<"The Factorial is: "<<endl;
-	for(int i=0; i<digit.size(); i++){
+	for(size_t i=0; i<digit.size(); i++){
 		cout<<digit[i];
 	}
 	cout<<endl;
